Reichelt canonizer returns http scheme again
HTTPS is used to hint the agent that a secure connection should be used.
As this is the desired standard now there is no benefit in adding it to
the URI scheme (plus all the schemes will change AGAIN should something
like TLS or an integrated security layer (HTTP 2.0?) come in the future.

That's why we stick with the http scheme and do the conversion to https
during the extraction only.

The Reichelt web server currently returns 301 with a new URL on the https
scheme when http URIs are resolved directly.